server returns Error in unhandled, example
