+++ client/client-1.go
@@ -71,8 +71,6 @@ func main() {
 		for _, d := range s.Details() {
 			switch info := d.(type) {
 			case *epb.QuotaFailure:
-				log.Printf("SENTRY...")
-				sentry.CaptureException(err) // s.Err()
 				log.Printf("Quota failure: %s", info)
 			default:
 				log.Printf("Unexpected type: %s", info)
