+++ client/client-2.go
@@ -54,6 +54,6 @@ func main() {
 	if err != nil {
 		log.Fatalf("could not greet: %v", err)
 	}
-	log.Printf(r.GetMessage())
-	// log.Printf("Greeting: %s", r.GetMessage())
+
+	log.Printf("Greeting: %s", r.GetMessage())
 }
